# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: EncadrePro API – gestion d'un atelier d'encadrement
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /produits:
    get:
      operationId: listProduits
      tags: [produits]
      summary: List products
      parameters:
        - name: categorie
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produit"

    post:
      operationId: createProduit
      tags: [produits]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduitBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
  /produits/{id}:
    put:
      operationId: updateProduit
      tags: [produits]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduitBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"

    delete:
      operationId: deleteProduit
      tags: [produits]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /produits/{id}/toggle:
    patch:
      operationId: toggleProduitActif
      tags: [produits]
      summary: Toggle product active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"

  # ──────────────── DEVIS ────────────────
...
  /factures:
    get:
      operationId: listFactures
      tags: [factures]
      summary: List invoices
      parameters:
        - name: statut
          in: query
          schema:
            type: string
        - name: client_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Facture"

    post:
      operationId: createFacture
      tags: [factures]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFactureBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"
  /factures/{id}:
    get:
      operationId: getFacture
      tags: [factures]
      summary: Get an invoice with lines and payments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactureDetail"

    delete:
      operationId: deleteFacture
      tags: [factures]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /factures/{id}/statut:
    patch:
      operationId: updateFactureStatut
      tags: [factures]
      summary: Update invoice status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatutBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"
  /factures/{id}/paiements:
    post:
      operationId: addPaiement
      tags: [factures]
      summary: Add a payment to an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaiementBody"
      responses:
        "201":
          description: Added payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acompte"

  # ──────────────── ATELIER ────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── DASHBOARD ──
...
    Produit:
      type: object
      properties:
        id: { type: integer }
        reference: { type: string, nullable: true }
        designation: { type: string }
        categorie: { type: string }
        unite_calcul: { type: string }
        prix_ht: { type: number }
        taux_tva: { type: number }
        actif: { type: integer }
        notes: { type: string, nullable: true }
        cree_le: { type: string }
      required: [id, designation, categorie, unite_calcul, prix_ht, taux_tva, actif, cree_le]
    CreateProduitBody:
      type: object
      properties:
        reference: { type: string, nullable: true }
        designation: { type: string }
        categorie: { type: string }
        unite_calcul: { type: string }
        prix_ht: { type: number }
        taux_tva: { type: number }
        notes: { type: string, nullable: true }
      required: [designation, categorie, unite_calcul, prix_ht, taux_tva]

    # ── DEVIS ──
    LigneDevis:
      type: object
      properties:
        id: { type: integer }
        devis_id: { type: integer }
        produit_id: { type: integer, nullable: true }
        designation: { type: string }
        unite_calcul: { type: string }
        largeur_m: { type: number, nullable: true }
        hauteur_m: { type: number, nullable: true }
        quantite: { type: number }
        quantite_calculee: { type: number, nullable: true }
        prix_unitaire_ht: { type: number }
        taux_tva: { type: number }
        total_ht: { type: number }
        total_ttc: { type: number }
        ordre: { type: integer }
      required: [id, devis_id, designation, unite_calcul, quantite, prix_unitaire_ht, taux_tva, total_ht, total_ttc, ordre]
...
    UpdateStatutBody:
      type: object
      properties:
        statut: { type: string }
      required: [statut]

    # ── FACTURES ──
    Acompte:
      type: object
      properties:
        id: { type: integer }
        facture_id: { type: integer }
        montant: { type: number }
        date_paiement: { type: string }
        mode_paiement: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        cree_le: { type: string }
      required: [id, facture_id, montant, date_paiement, cree_le]
    Facture:
      type: object
      properties:
        id: { type: integer }
        numero: { type: string }
        devis_id: { type: integer, nullable: true }
        client_id: { type: integer }
        client_nom: { type: string, nullable: true }
        client_prenom: { type: string, nullable: true }
        date_creation: { type: string }
        date_echeance: { type: string, nullable: true }
        statut: { type: string }
        sous_total_ht: { type: number }
        total_tva_10: { type: number }
        total_tva_20: { type: number }
        total_ttc: { type: number }
        total_paye: { type: number }
        solde_restant: { type: number }
        notes: { type: string, nullable: true }
        conditions: { type: string, nullable: true }
        cree_le: { type: string }
        modifie_le: { type: string }
      required: [id, numero, client_id, date_creation, statut, sous_total_ht, total_tva_10, total_tva_20, total_ttc, total_paye, solde_restant, cree_le, modifie_le]
    FactureDetail:
      allOf:
        - $ref: "#/components/schemas/Facture"
        - type: object
          properties:
            lignes:
              type: array
              items:
                $ref: "#/components/schemas/LigneDevis"
            paiements:
              type: array
              items:
                $ref: "#/components/schemas/Acompte"
          required: [lignes, paiements]
    CreateFactureBody:
      type: object
      properties:
        client_id: { type: integer }
        notes: { type: string, nullable: true }
      required: [client_id]
    CreatePaiementBody:
      type: object
      properties:
        montant: { type: number }
        date_paiement: { type: string }
        mode_paiement: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      required: [montant, date_paiement]

    # ── ATELIER ──
...