# lib/api-zod/src/generated/types/acompte.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EncadrePro API – gestion d'un atelier d'encadrement
 * OpenAPI spec version: 1.0.0
 */

export interface Acompte {
  id: number;
  facture_id: number;
  montant: number;
  date_paiement: string;
  mode_paiement?: string | null;
  notes?: string | null;
  cree_le: string;
}