# lib/api-zod/src/generated/types/factureDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EncadrePro API – gestion d'un atelier d'encadrement
 * OpenAPI spec version: 1.0.0
 */
import type { Acompte } from "./acompte";
import type { Facture } from "./facture";
import type { LigneDevis } from "./ligneDevis";

export type FactureDetail = Facture & {
  lignes: LigneDevis[];
  paiements: Acompte[];
};